+++ packages/cmf/__tests__/middlewares/http.test.js
@@ -13,10 +13,7 @@ import {
 } from '../../src/middlewares/http/middleware';
 import interceptors from '../../src/httpInterceptors';
 
-import {
-	HTTP_METHODS,
-	HTTP_STATUS,
-} from '../../src/middlewares/http/constants';
+import { HTTP_METHODS, HTTP_STATUS } from '../../src/middlewares/http/constants';
 
 describe('CMF http middleware', () => {
 	beforeEach(() => {
@@ -94,9 +91,9 @@ describe('CMF http middleware', () => {
 		const httpConfig = {
 			headers: {
 				'Accept-Language': 'fr-FR',
-			}
+			},
 		};
-		const defaultOptions = { url: '/url1', headers: { Accept: 'application/json'} };
+		const defaultOptions = { url: '/url1', headers: { Accept: 'application/json' } };
 		const options = mergeConfiguredHeader(httpConfig)(defaultOptions);
 		expect(options.url).toBe('/url1');
 		expect(options.headers.Accept).toBe('application/json');
@@ -108,11 +105,11 @@ describe('CMF http middleware', () => {
 		const httpConfig = {
 			headers: {
 				'Accept-Language': 'fr-FR',
-			}
+			},
 		};
 		const defaultOptions = {
 			url: '/url1',
-			headers: { Accept: 'application/json', 'Accept-Language': 'en-US', }
+			headers: { Accept: 'application/json', 'Accept-Language': 'en-US' },
 		};
 		const options = mergeConfiguredHeader(httpConfig)(defaultOptions);
 		expect(options.url).toBe('/url1');
@@ -125,14 +122,14 @@ describe('CMF http middleware', () => {
 		const httpConfig = {
 			headers: {
 				'Accept-Language': 'fr-FR',
-			}
+			},
 		};
 		const actionOptions = {
 			url: '/url1',
 			headers: {
 				Accept: 'application/json',
 				'Content-Type': 'a content type',
-			}
+			},
 		};
 		const options = mergeConfiguredHeader(httpConfig)(actionOptions);
 		expect(options.url).toBe('/url1');
@@ -145,7 +142,7 @@ describe('CMF http middleware', () => {
 		const httpConfig = {
 			headers: {
 				'Accept-Language': 'fr-FR',
-			}
+			},
 		};
 		const actionOptions = {
 			url: '/url1',
@@ -153,7 +150,7 @@ describe('CMF http middleware', () => {
 			headers: {
 				Accept: 'application/json',
 				'Content-Type': 'a content type',
-			}
+			},
 		};
 		const options = mergeConfiguredHeader(httpConfig)(actionOptions);
 		expect(options.url).toBe('/url1');
@@ -500,7 +497,6 @@ describe('httpMiddleware configuration', () => {
 		const middleware = httpMiddleware(httpDefaultConfig)(store)(next);
 		expect(typeof middleware).toBe('function');
 		middleware(action).then(() => {
-
 			// then
 			const firstCall = global.fetch.mock.calls[0];
 			const firstCallSecondParam = firstCall[1];
@@ -508,10 +504,7 @@ describe('httpMiddleware configuration', () => {
 			expect(firstCallSecondParam).toHaveProperty('body', expectedBody);
 			expect(firstCallSecondParam).toHaveProperty('credentials', expectedCredentials);
 			expect(firstCallSecondParam).toHaveProperty('headers.Accept', expectedAccept);
-			expect(firstCallSecondParam).toHaveProperty(
-				'headers.Content-Type',
-				expectedContentType,
-			);
+			expect(firstCallSecondParam).toHaveProperty('headers.Content-Type', expectedContentType);
 			expect(firstCallSecondParam).toHaveProperty('headers.headerKey', expectedCSRFKeyValue);
 			expect(firstCallSecondParam).toHaveProperty('method', expectedMethod);
 			expect(firstCallSecondParam).toHaveProperty('onError', expectedOnError);
@@ -578,14 +571,8 @@ describe('httpMiddleware configuration', () => {
 			expect(firstCallSecondParam).toHaveProperty('body', expectedBody);
 			expect(firstCallSecondParam).toHaveProperty('credentials', expectedCredentials);
 			expect(firstCallSecondParam).toHaveProperty('headers.Accept', expectedAccept);
-			expect(firstCallSecondParam).toHaveProperty(
-				'headers.Content-Type',
-				expectedContentType,
-			);
-			expect(firstCallSecondParam).toHaveProperty(
-				'headers.X-CSRF-Token',
-				expectedCSRFKeyValue,
-			);
+			expect(firstCallSecondParam).toHaveProperty('headers.Content-Type', expectedContentType);
+			expect(firstCallSecondParam).toHaveProperty('headers.X-CSRF-Token', expectedCSRFKeyValue);
 			expect(firstCallSecondParam).toHaveProperty('method', expectedMethod);
 			expect(firstCallSecondParam).toHaveProperty('onError', expectedOnError);
 			expect(firstCallSecondParam).toHaveProperty('onResponse', expectedOnResponse);
@@ -638,7 +625,7 @@ describe('httpMiddleware configuration', () => {
 			expect(interceptor.request).toHaveBeenCalled();
 			const augmentedConfig = interceptor.request.mock.calls[0][0];
 			expect(augmentedConfig.url).toBe(action.url);
-			expect(interceptor.response).toHaveBeenCalledWith({ data: response, headers: {}});
+			expect(interceptor.response).toHaveBeenCalledWith({ data: response, headers: {} });
 			// eslint-disable-next-line no-underscore-dangle
 			interceptors._clear();
 			done();