+++ packages/design-system/.storybook/preview.js
@@ -1,23 +1,23 @@
 import React from 'react';
 import { Helmet } from 'react-helmet';
 import { I18nextProvider } from 'react-i18next';
+import { useLocalStorage } from 'react-use';
+
 import prettier from 'prettier/standalone';
 import prettierBabel from 'prettier/parser-babel';
 
 import { addons } from '@storybook/addons';
+
 import { DocsContainer } from '@storybook/addon-docs';
-import { UPDATE_GLOBALS } from '@storybook/core-events';
+import { UPDATE_GLOBALS, SET_STORIES } from '@storybook/core-events';
 import { TableOfContents, BackToTop } from 'storybook-docs-toc';
-import { useLocalStorage } from 'react-use';
 
 import 'focus-outline-manager';
 
 import i18n from './i18n';
 
-import Divider from '../src/components/Divider';
-import Form from '../src/components/Form';
-import ThemeProvider from '../src/components/ThemeProvider';
-import { IconsProvider } from '../src/components/IconsProvider';
+import { BadgeFigma, BadgeStorybook, BadgeReact, BadgeI18n, Badges } from './docs';
+import { Divider, Form, IconsProvider, ThemeProvider } from '../src';
 
 import { light, dark } from '../src/themes';
 
@@ -57,17 +57,17 @@ const StorybookGlobalStyle = ThemeProvider.createGlobalStyle(
 	.sb-show-main.sb-main-padded {
 		padding: 0;
 	}
-	
+
 	.sbdocs.sbdocs-preview {
 		color: ${theme?.colors.textColor};
 		background: ${theme?.colors.backgroundColor};
 	}
-	
+
 	.sbdocs .figma-iframe--light {
 		position: ${theme?.id === 'light' && hasFigmaIframe ? 'relative' : 'absolute'};
 		left:  ${theme?.id === 'light' && hasFigmaIframe ? 'auto' : '-9999rem'};
 	}
-	
+
 	.sbdocs .figma-iframe--dark {
 		position: ${theme?.id === 'dark' && hasFigmaIframe ? 'relative' : 'absolute'};
 		left:  ${theme?.id === 'dark' && hasFigmaIframe ? 'auto' : '-9999rem'};
@@ -76,6 +76,27 @@ const StorybookGlobalStyle = ThemeProvider.createGlobalStyle(
 );
 
 const channel = addons.getChannel();
+
+let statusByPage = {};
+channel.once(SET_STORIES, eventData => {
+	statusByPage = Object.entries(eventData.stories).reduce(
+		(acc, [name, { title, componentId, parameters }]) => {
+			['components', 'templates', 'pages'].forEach(prefix => {
+				if (name.toLocaleLowerCase().startsWith(prefix)) {
+					if (!acc[componentId]) {
+						acc[componentId] = {
+							title,
+							componentId,
+							parameters,
+						};
+					}
+				}
+			});
+			return acc;
+		},
+		{},
+	);
+});
 
 export const parameters = {
 	docs: {
@@ -87,6 +108,8 @@ export const parameters = {
 				true,
 			);
 
+			const { id, parameters, globals, title } = props.context;
+
 			React.useEffect(() => {
 				channel.emit(UPDATE_GLOBALS, {
 					globals: { theme: hasDarkMode ? 'dark' : 'light' },
@@ -94,12 +117,16 @@ export const parameters = {
 			}, [hasDarkMode]);
 
 			React.useEffect(() => {
-				const theme = props.context.globals?.theme;
+				channel.emit('SET_STATUSES_BY_PAGE', statusByPage);
+			}, [statusByPage]);
+
+			const { theme } = globals;
+			React.useEffect(() => {
 				const hasDarkModeFromToolbar = theme === 'dark';
 				if (hasDarkModeFromToolbar != hasDarkMode) {
 					setDarkMode(hasDarkModeFromToolbar);
 				}
-			}, [props.context.globals?.theme]);
+			}, [theme]);
 
 			React.useEffect(() => {
 				document
@@ -107,11 +134,24 @@ export const parameters = {
 					.forEach(link => (link.disabled = !hasBootstrapStylesheet));
 			}, [hasBootstrapStylesheet]);
 
-			const title = props.context.title;
 			const titleArray = title?.split('/');
 
 			const docsTitle = title?.replaceAll(/\//gi, ' / ');
 			const docsCategory = titleArray[0];
+
+			const { status = {}, figmaLink } = parameters;
+
+			const githubLink =
+				'https://github.com/Talend/ui/tree/master/packages/design-system/' +
+				parameters.fileName
+					.split('/')
+					.slice(1, parameters.fileName.split('/').length - 1)
+					.join('/')
+					.replace('/docs', '');
+
+			const isDesignSystemElementPage = ['components/', 'templates/', 'pages/'].find(term =>
+				title?.toLocaleLowerCase().startsWith(term),
+			);
 
 			return (
 				<>
@@ -119,26 +159,18 @@ export const parameters = {
 						<title>{docsTitle}</title>
 						<meta property="og:title" content={titleArray[titleArray.length - 1]} />
 						<meta property="og:type" content="article" />
-						<meta
-							property="og:url"
-							content={`https://design.talend.com/?path=/docs/${props.context.id}`}
-						/>
+						<meta property="og:url" content={`https://design.talend.com/?path=/docs/${id}`} />
 						<meta
 							property="og:image"
 							content={`https://via.placeholder.com/1000x500/F3F3F3/FF6D70?text=${docsTitle}`}
 						/>
 						{titleArray.length > 1 && <meta property="article:section" content={docsCategory} />}
 					</Helmet>
+
 					<IconsProvider bundles={['https://unpkg.com/@talend/icons/dist/svg-bundle/all.svg']} />
-					<ThemeProvider theme={hasDarkMode ? dark : light}>
-						<ThemeProvider.GlobalStyle />
-						<StorybookGlobalStyle hasFigmaIframe={hasFigmaIframe} />
-					</ThemeProvider>
 					<TableOfContents>
-						{['component', 'template', 'page'].find(term =>
-							docsCategory.toLocaleLowerCase().includes(term),
-						) && (
-							<ThemeProvider>
+						{isDesignSystemElementPage && (
+							<ThemeProvider theme={light}>
 								<Divider />
 								<Form.Switch
 									label={'Dark mode'}
@@ -162,7 +194,24 @@ export const parameters = {
 							</ThemeProvider>
 						)}
 					</TableOfContents>
-					<DocsContainer {...props} />
+
+					{isDesignSystemElementPage && status && (
+						<Badges>
+							<BadgeFigma status={status.figma} href={figmaLink} />
+							<BadgeStorybook status={status.storybook} />
+							<BadgeReact status={status.react} href={githubLink} />
+							<BadgeI18n status={status.i18n} />
+						</Badges>
+					)}
+
+					<I18nextProvider i18n={i18n}>
+						<ThemeProvider theme={hasDarkMode ? dark : light}>
+							<ThemeProvider.GlobalStyle />
+							<StorybookGlobalStyle hasFigmaIframe={hasFigmaIframe} />
+							<DocsContainer {...props} />
+						</ThemeProvider>
+					</I18nextProvider>
+
 					<BackToTop />
 				</>
 			);
@@ -208,6 +257,7 @@ export const parameters = {
 				[
 					'Accordion',
 					'Button',
+					'ButtonIcon',
 					'Combobox',
 					'Divider',
 					'Drawer',
@@ -235,18 +285,25 @@ export const parameters = {
 
 export const decorators = [
 	(Story, context) => {
-		i18n.changeLanguage(context.globals?.locale);
-		const theme = getTheme(context.globals?.theme);
-		return (
+		const { globals = {}, viewMode } = context;
+
+		const { locale: localeKey, theme: themeKey } = globals;
+		if (localeKey) i18n.changeLanguage(localeKey);
+		const theme = getTheme(themeKey);
+
+		const themedStory = (
+			<ThemeProvider theme={theme}>
+				<ThemeProvider.GlobalStyle />
+				<Story {...context} />
+			</ThemeProvider>
+		);
+
+		return viewMode === 'docs' ? (
+			themedStory
+		) : (
 			<I18nextProvider i18n={i18n}>
 				<IconsProvider bundles={['https://unpkg.com/@talend/icons/dist/svg-bundle/all.svg']} />
-				<ThemeProvider theme={theme}>
-					<ThemeProvider.GlobalStyle />
-					<StorybookGlobalStyle />
-					<React.Suspense fallback={null}>
-						<Story {...context} />
-					</React.Suspense>
-				</ThemeProvider>
+				{themedStory}
 			</I18nextProvider>
 		);
 	},