+++ packages/containers/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @talend/react-containers
 
+## 7.0.1
+
+### Patch Changes
+
+- 618951c8b: chore(deps): auto update for maintenance purpose
+- Updated dependencies [618951c8b]
+- Updated dependencies [f5e3a7ead]
+  - @talend/react-cmf@7.0.1
+  - @talend/react-components@6.43.5
+  - @talend/react-forms@7.0.1
+
 ## 7.0.0
 
 ### Major Changes